# swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "App Craw Data",
      "description": "",
      "contact": {
        "email": "khaiquangletn@gmail.com"
      },
      "version": ""
    },
    "servers": [
      {
        "url": "http://localhost:3000/api/v1",
        "description": ""
      }  
    ],
    "tags": [
      {
        "name": "data",
        "description": "Everything about your Datas"
      }   
    ],
    "paths": {
      "/giaoduc/{id}": {
        "get": {
          "tags": [
            "Giáo dục"
          ],
          "summary": "Find article by ID",
          "description": "Returns a single article",
          "operationId": "getArticleById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of article to return ( 1 - 30 )",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Giaoduc"
                  }
                }          
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Article not found"
            }
          }
        }
      },
      "/tretho/{id}": {
        "get": {
          "tags": [
            "Trẻ thơ"
          ],
          "summary": "Find article by ID",
          "description": "Returns a single article",
          "operationId": "getArticleById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of article to return ( 1 - 20 )",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Tretho"
                  }
                }          
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Article not found"
            }
          }
        }
      },
      "/findByType": {
        "get": {
          "tags": [
            "Trẻ thơ"
          ],
          "summary": "Find article by Type",
          "description": "Returns multiple  article",
          "operationId": "findArticlesByType",
          "parameters": [
            {
              "name": "type",
              "in": "query",
              "description": "",
              "required": false,
              "explode": true,
              "schema": {
                "type": "string",
                "default": "available",
                "enum": [
                  "Ngẫm chuyện đời",
                  "Trải nghiệm cuộc sống",
                  "Bí kíp du lịch",
                  "Du lịch nước ngoài"
                ]
              }
            }
          ],
          "responses": {
            "200": {
              "description":"successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Tretho"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid status value"
            }
          }
          }
        } 
    },
    
    "components": {
      "schemas": {
        "Giaoduc": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "title": {
              "type": "string"
            },
            "content": {
              "type": "string"
            },
            "url": {
              "type": "string"
            },
            "time": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Tretho": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "title": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "url": {
              "type": "string"
            },
            "time": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "ApiResponse": {
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "type": {
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          }
        }
      }
    }
  }